+++ src/pages/components/EmptyOrdersComponent/index.tsx
@@ -1,8 +1,10 @@
+import { FC } from 'react';
 import { Typography } from '@mui/material';
 import { t } from 'i18next';
+
 import { emptyOrdersPageStyles } from './styles';
 
-const EmptyOrdersComponent = () => {
+const EmptyOrdersComponent: FC = () => {
   return (
     <Typography sx={emptyOrdersPageStyles}>
       {t('orderManagement.thereIsNoOrders')}
